Initial code for frontend data speedups.
Added a package called "miro.data" which the frontend will use to fetch its
data.  I choose a generic name because, who knows, maybe the backend will want
to use this too.

Added the ItemTracker class.  It uses SQL to track changes to an item list.
Added some basic tests for it.