Sort/filter using the remote_downloader and dlstats table in ItemTracker
Added code to populate the dlstats table when we get updates from the
downloader.

Added code to tell when we need to reload the list for ItemLists that sort on
downloader / dlstats columns.  Since this is basically all the time that the
downloader changes, we just use a flag rather than figuring out which columns
actually changed.