Modified itemlist module so that we can share ItemList objects
This is to handle the situation where the playback code and the
ItemListController are working on the same tab.  For example, I play an item,
switch tabs, then switch back tabs.  At that point, we want the
ItemListController and the playback code to share the same ItemList.  This is
good because:

- We don't have to process the ItemsChanged message twice.
- If a user changes the filters, the search, etc, the changes will be
  reflected in the playback.

Created an ItemListPool class to handle this.  It also handles calling
on_items_changed() for all active ItemLists.  Also refactored ItemList a bit.